test: fold deadlock tests into the general suites
The sync and async self-deadlock tests lived in dedicated
test_self_deadlock.py / test_async_self_deadlock.py files. Move them into
test_filelock.py and test_async_filelock.py next to the rest of each
flavor's coverage, matching how the other behaviours are organised, and
drop the now-empty files. Names are unchanged except test_cleanup_on_release
becomes test_deadlock_registry_cleanup_on_release, which is no longer
self-evident once it sits among unrelated release tests.
