+++ tests/test_async_filelock.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import asyncio
 import gc
 import logging
+import sys
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path, PurePath
 
@@ -355,3 +356,137 @@ def test_del_after_loop_close_does_not_raise(lock_type: type[BaseAsyncFileLock],
 
     with ThreadPoolExecutor(max_workers=1) as pool:
         pool.submit(_run).result(timeout=10)
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_same_task_different_instances_raises(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path)
+    async with lock1:
+        lock2 = lock_type(lock_path)
+        with pytest.raises(RuntimeError, match="Deadlock"):
+            await lock2.acquire()
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_finite_timeout_gives_timeout_not_deadlock(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path)
+    async with lock1:
+        lock2 = lock_type(lock_path, timeout=0.1)
+        with pytest.raises(Timeout):
+            await lock2.acquire()
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_non_blocking_gives_timeout_not_deadlock(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path)
+    async with lock1:
+        lock2 = lock_type(lock_path, blocking=False)
+        with pytest.raises(Timeout):
+            await lock2.acquire()
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_different_paths_no_conflict(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock1 = lock_type(tmp_path / "a.lock")
+    lock2 = lock_type(tmp_path / "b.lock")
+    async with lock1, lock2:
+        assert lock1.is_locked
+        assert lock2.is_locked
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_same_instance_reentrant_works(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock = lock_type(tmp_path / "test.lock")
+    async with lock:
+        async with lock:
+            assert lock.is_locked
+        assert lock.is_locked
+    assert not lock.is_locked
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_singleton_avoids_deadlock(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path, is_singleton=True)
+    async with lock1:
+        lock2 = lock_type(lock_path, is_singleton=True)
+        assert lock1 is lock2
+        async with lock2:
+            assert lock2.is_locked
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_different_tasks_no_false_positive(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    # The registry is per-thread; a second acquire from a different task (asyncio's counterpart to the sync
+    # test's separate thread) must not be mistaken for a reentrant deadlock.
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path, timeout=0)
+    await lock1.acquire()
+
+    error: BaseException | None = None
+
+    async def acquire_other() -> None:
+        nonlocal error
+        lock2 = lock_type(lock_path, timeout=0)
+        try:
+            await lock2.acquire()
+        except BaseException as exc:
+            error = exc
+
+    await asyncio.create_task(acquire_other())
+    await lock1.release()
+
+    assert not isinstance(error, RuntimeError), "Should not raise RuntimeError in a different task"
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="unix-only symlink test")
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_symlink_same_canonical_path(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    symlink_path = tmp_path / "link.lock"
+    symlink_path.symlink_to(lock_path)
+
+    lock1 = lock_type(lock_path)
+    async with lock1:
+        lock2 = lock_type(symlink_path)
+        with pytest.raises(RuntimeError, match="Deadlock"):
+            await lock2.acquire()
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_deadlock_registry_cleanup_on_release(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path)
+    await lock1.acquire()
+    await lock1.release()
+
+    lock2 = lock_type(lock_path)
+    async with lock2:
+        assert lock2.is_locked
+
+
+@pytest.mark.parametrize("lock_type", [AsyncFileLock, AsyncSoftFileLock])
+@pytest.mark.asyncio
+async def test_force_release_clears_registry(tmp_path: Path, lock_type: type[BaseAsyncFileLock]) -> None:
+    lock_path = tmp_path / "test.lock"
+    lock1 = lock_type(lock_path)
+    async with lock1:
+        await lock1.acquire()
+        assert lock1.lock_counter == 2
+    await lock1.release(force=True)
+
+    lock2 = lock_type(lock_path)
+    async with lock2:
+        assert lock2.is_locked
